+++ cmd/kyfinfo/main.go
@@ -1,0 +1,10 @@
+package main
+
+/*
+Copyright © 2025 NAME HERE <EMAIL ADDRESS>
+*/
+import "github.com/puppetma4ster/koyane-framework/cmd/kyfinfo/app"
+
+func main() {
+	app.Execute()
+}