+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## upcoming
 - added permutation wordlist generation
+- added piping with stdout for oermutation
+- added quiet start mode with parameter "-q"
 - hashcat potfile extraction is now in generate instead of edit
+- Fixed a bug where “--extract-hc-potfile” printed the output starting from the first “:”
+- Change: Print duplicate words only with flag "-d"
+- Split the program commands into individual CLI versions kyfdb,kyfgen,kyfinfo and kyf edit
 
 ## koyane-framework 0.7.0-beta
 - Complete program remastered in GO